Compila todo pero hay error en el login (socket NULL)

+++ rvr_repo/Proyecto/Vector2D.h
@@ -13,8 +13,8 @@ public:
 	Vector2D() {}
 	Vector2D(double x, double y) :
 		x(x), y(y) {}
-	int getX() const;
-	int getY() const;
+	uint16_t getX() const;
+	uint16_t getY() const;
 	void normaliza();
 	float magnitude();
 	Vector2D operator+(const Vector2D& v) const;
